+++ event/stream/decode_test.go
@@ -1,6 +1,7 @@
 package stream
 
 import (
+	"strings"
 	"testing"
 	"time"
 
@@ -246,4 +247,104 @@ func TestDecode_StateValueIsCaseInsensitive(t *testing.T) {
 			assert.Equal(t, tc.want, ev.State)
 		})
 	}
+}
+
+// --- Edge cases for state extraction and time parsing ----------------
+
+func TestDecode_PropertyOperationIsCaseSensitive(t *testing.T) {
+	// Per WS-Notification §3.3 PropertyOperation values are
+	// 'Initialized' / 'Changed' / 'Deleted'. Lowercased forms are
+	// malformed and should fall through to PropertyUnknown.
+	in := msg("tns1:VideoSource/MotionAlarm", "changed", "", nil, nil)
+	ev := decode(in, "dev", time.Now())
+	assert.Equal(t, PropertyUnknown, ev.Operation)
+}
+
+func TestDecode_StateValueTrimsWhitespace(t *testing.T) {
+	tests := []struct {
+		name  string
+		value string
+		want  State
+	}{
+		{"leading_trailing", "  true  ", StateActive},
+		{"tab_newline", "\ttrue\n", StateActive},
+		{"only_spaces", "   ", StateUnknown},
+	}
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			in := msg("tns1:VideoSource/MotionAlarm", "Changed", "",
+				nil, map[string]string{"State": tc.value})
+			ev := decode(in, "dev", time.Now())
+			assert.Equal(t, tc.want, ev.State)
+		})
+	}
+}
+
+func TestDecode_SimpleItemEmptyValueIsUnknownState(t *testing.T) {
+	in := msg("tns1:VideoSource/MotionAlarm", "Changed", "",
+		nil, map[string]string{"State": ""})
+	ev := decode(in, "dev", time.Now())
+	assert.Equal(t, StateUnknown, ev.State)
+	v, ok := ev.Data["State"]
+	assert.True(t, ok)
+	assert.Equal(t, "", v)
+}
+
+func TestDecode_DeviceTimeAdditionalLayouts(t *testing.T) {
+	tests := []struct {
+		name string
+		in   string
+		want time.Time
+	}{
+		{"compact_offset", "2026-05-21T12:30:00+0200", time.Date(2026, 5, 21, 10, 30, 0, 0, time.UTC)},
+		{"compact_offset_subsec", "2026-05-21T12:30:00.500+0200", time.Date(2026, 5, 21, 10, 30, 0, 500_000_000, time.UTC)},
+		{"naked_no_tz", "2026-05-21T10:30:00", time.Date(2026, 5, 21, 10, 30, 0, 0, time.UTC)},
+	}
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			in := msg("tns1:VideoSource/MotionAlarm", "Changed", tc.in, nil, nil)
+			ev := decode(in, "dev", time.Now())
+			assert.True(t, ev.DeviceTime.Equal(tc.want),
+				"input=%q got=%v want=%v", tc.in, ev.DeviceTime, tc.want)
+		})
+	}
+}
+
+func TestDecode_DeviceTimeStillRejectsNonsense(t *testing.T) {
+	for _, s := range []string{"hello", "2026-13-45T99:99:99", strings.Repeat("9", 50)} {
+		in := msg("tns1:VideoSource/MotionAlarm", "Changed", s, nil, nil)
+		ev := decode(in, "dev", time.Now())
+		assert.True(t, ev.DeviceTime.IsZero(), "input=%q should yield zero, got %v", s, ev.DeviceTime)
+	}
+}
+
+// --- First-boolean-wins with explicit slice order --------------------
+
+type pair struct{ k, v string }
+
+func simpleItemsFromPairs(pairs []pair) []event.SimpleItem {
+	out := make([]event.SimpleItem, len(pairs))
+	for i, p := range pairs {
+		out[i] = event.SimpleItem{
+			Name:  xsd.AnyType(p.k),
+			Value: xsd.AnyType(p.v),
+		}
+	}
+	return out
+}
+
+func TestExtractState_FirstBooleanLikeWins(t *testing.T) {
+	// Documented behaviour: when multiple Data items have boolean-like
+	// values, the first by slice order wins. Use explicit slice
+	// construction so the assertion does not depend on map iteration
+	// order.
+	in := msg("tns1:VideoSource/MotionAlarm", "Changed", "", nil, nil)
+	in.Message.Message.Data.SimpleItem = simpleItemsFromPairs([]pair{
+		{"ObjectId", "42"},
+		{"State", "true"},
+		{"Trailer", "false"},
+	})
+	ev := decode(in, "dev", time.Now())
+	assert.Equal(t, StateActive, ev.State,
+		"first boolean-like value (State=true) must win, not Trailer=false")
 }
