+++ event/stream/reconnect_test.go
@@ -29,15 +29,13 @@ const createPullPointRespAlt = `<?xml version="1.0" encoding="UTF-8"?>
   </env:Body>
 </env:Envelope>`
 
+// --- Recreate after pull failures ------------------------------------
+
 func TestStream_RecreatesSubscriptionAfterRepeatedPullErrors(t *testing.T) {
 	fc := newFakeCaller()
-	// Initial subscription.
 	fc.queueCallMethod(createPullPointResp, nil)
-	// Recreated subscription returns a *different* endpoint.
 	fc.queueCallMethod(createPullPointRespAlt, nil)
 
-	// First pull fails. With ReconnectAfterFailures=1 this triggers a
-	// recreate; subsequent pulls go to PullSub_2 which we'll observe.
 	fc.queueSendSoap("", errors.New("transient failure"))
 	fc.queueSendSoap(pullMessagesResp(motionMsg("true")), nil)
 
@@ -48,7 +46,7 @@ func TestStream_RecreatesSubscriptionAfterRepeatedPullErrors(t *testing.T) {
 		PullTimeout:            50 * time.Millisecond,
 		ReconnectAfterFailures: 1,
 		RetryBackoff:           10 * time.Millisecond,
-		InitialTermination:     30 * time.Second, // keep renew quiet
+		InitialTermination:     30 * time.Second,
 	})
 	require.NoError(t, err)
 	defer s.Close()
@@ -60,8 +58,6 @@ func TestStream_RecreatesSubscriptionAfterRepeatedPullErrors(t *testing.T) {
 	defer fc.mu.Unlock()
 	require.Len(t, fc.callMethodCalls, 2,
 		"expected exactly 2 CallMethod calls (initial + recreate)")
-	// The PullMessages call that delivered the motion event must
-	// target the new endpoint.
 	var newEndpointPulls int
 	for _, c := range fc.sendSoapCalls {
 		if c[0] == "http://camera.local/onvif/Events/PullSub_2" {
@@ -75,9 +71,6 @@ func TestStream_RecreatesSubscriptionAfterRepeatedPullErrors(t *testing.T) {
 func TestStream_BackoffWhenRecreateFails(t *testing.T) {
 	fc := newFakeCaller()
 	fc.queueCallMethod(createPullPointResp, nil)
-	// After the initial successful create, every CallMethod (recreate)
-	// and SendSoap (pull) fails. The loop should keep retrying with
-	// exponential backoff rather than blocking forever or spinning.
 	fc.mu.Lock()
 	fc.defaultCall = fakeResp{err: errors.New("recreate fail")}
 	fc.defaultSendSoap = fakeResp{err: errors.New("pull fail")}
@@ -117,4 +110,206 @@ func TestStream_ReconnectAfterFailuresDefault(t *testing.T) {
 func TestStream_RetryBackoffDefault(t *testing.T) {
 	o := defaultOptions()
 	assert.Equal(t, time.Second, o.RetryBackoff)
+}
+
+func TestStream_DisableReconnectKeepsRetryingOriginalEndpoint(t *testing.T) {
+	fc := newFakeCaller()
+	fc.queueCallMethod(createPullPointResp, nil)
+	fc.mu.Lock()
+	fc.defaultSendSoap = fakeResp{err: errors.New("pull fail")}
+	fc.mu.Unlock()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	s, err := newStream(ctx, fc, Options{
+		PullTimeout:        10 * time.Millisecond,
+		RetryBackoff:       10 * time.Millisecond,
+		InitialTermination: 30 * time.Second,
+		DisableReconnect:   true,
+	})
+	require.NoError(t, err)
+	defer s.Close()
+
+	time.Sleep(200 * time.Millisecond)
+	fc.mu.Lock()
+	calls := len(fc.callMethodCalls)
+	fc.mu.Unlock()
+	assert.Equal(t, 1, calls, "DisableReconnect must prevent recreate; got %d CallMethod calls", calls)
+}
+
+func TestStream_RecreateResetsFailuresAndBackoffOnSuccess(t *testing.T) {
+	fc := newFakeCaller()
+	fc.queueCallMethod(createPullPointResp, nil)
+	fc.queueCallMethod(createPullPointRespAlt, nil)
+	fc.queueSendSoap("", errors.New("first failure"))
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	s, err := newStream(ctx, fc, Options{
+		PullTimeout:            10 * time.Millisecond,
+		ReconnectAfterFailures: 1,
+		RetryBackoff:           10 * time.Millisecond,
+		InitialTermination:     30 * time.Second,
+	})
+	require.NoError(t, err)
+	defer s.Close()
+
+	time.Sleep(200 * time.Millisecond)
+	fc.mu.Lock()
+	calls := len(fc.callMethodCalls)
+	fc.mu.Unlock()
+	assert.Equal(t, 2, calls,
+		"after one failure + successful recreate, no further recreates expected; got %d", calls)
+}
+
+func TestStream_PullPointMutationVisibleToRenewLoopUnderRace(t *testing.T) {
+	// Drives the pullPoint write-by-pullLoop / read-by-renewLoop race
+	// so -race actually exercises the mutex critical sections.
+	fc := newFakeCaller()
+	fc.queueCallMethod(createPullPointResp, nil)
+	for i := 0; i < 50; i++ {
+		fc.queueCallMethod(createPullPointRespAlt, nil)
+	}
+	fc.mu.Lock()
+	fc.defaultSendSoap = fakeResp{err: errors.New("recurring pull fail")}
+	fc.mu.Unlock()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	s, err := newStream(ctx, fc, Options{
+		PullTimeout:            5 * time.Millisecond,
+		ReconnectAfterFailures: 1,
+		RetryBackoff:           1 * time.Millisecond,
+		InitialTermination:     20 * time.Millisecond,
+		RenewMargin:            2 * time.Millisecond,
+	})
+	require.NoError(t, err)
+	defer s.Close()
+
+	time.Sleep(300 * time.Millisecond)
+}
+
+// --- Typed errors from the reconnect path ----------------------------
+
+func TestStream_PullErrorIsTypedErrPullFailed(t *testing.T) {
+	fc := newFakeCaller()
+	fc.queueCallMethod(createPullPointResp, nil)
+	fc.queueSendSoap("", errors.New("transient"))
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	s, err := newStream(ctx, fc, Options{
+		PullTimeout:        50 * time.Millisecond,
+		RetryBackoff:       10 * time.Millisecond,
+		InitialTermination: 30 * time.Second,
+	})
+	require.NoError(t, err)
+	defer s.Close()
+
+	select {
+	case e := <-s.Errors():
+		var pullErr ErrPullFailed
+		require.True(t, errors.As(e, &pullErr), "expected ErrPullFailed, got %T: %v", e, e)
+		assert.Contains(t, pullErr.Err.Error(), "transient")
+	case <-time.After(time.Second):
+		t.Fatal("expected ErrPullFailed on Errors channel")
+	}
+}
+
+func TestStream_RecreateErrorIsTypedErrRecreateFailed(t *testing.T) {
+	fc := newFakeCaller()
+	fc.queueCallMethod(createPullPointResp, nil)
+	fc.mu.Lock()
+	fc.defaultCall = fakeResp{err: errors.New("recreate fail")}
+	fc.defaultSendSoap = fakeResp{err: errors.New("pull fail")}
+	fc.mu.Unlock()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	s, err := newStream(ctx, fc, Options{
+		PullTimeout:            10 * time.Millisecond,
+		ReconnectAfterFailures: 1,
+		RetryBackoff:           10 * time.Millisecond,
+		InitialTermination:     30 * time.Second,
+	})
+	require.NoError(t, err)
+	defer s.Close()
+
+	deadline := time.Now().Add(time.Second)
+	var sawRecreate bool
+	for time.Now().Before(deadline) && !sawRecreate {
+		select {
+		case e := <-s.Errors():
+			var rec ErrRecreateFailed
+			if errors.As(e, &rec) {
+				sawRecreate = true
+				assert.Contains(t, rec.Err.Error(), "recreate fail")
+			}
+		case <-time.After(50 * time.Millisecond):
+		}
+	}
+	assert.True(t, sawRecreate, "expected at least one ErrRecreateFailed on Errors")
+}
+
+func TestStream_AfterReconnectFlagSetOnFirstPostRecreateBatch(t *testing.T) {
+	fc := newFakeCaller()
+	fc.queueCallMethod(createPullPointResp, nil)
+	fc.queueCallMethod(createPullPointRespAlt, nil)
+
+	fc.queueSendSoap("", errors.New("transient"))
+	fc.queueSendSoap(pullMessagesResp(motionMsg("true")), nil)
+	fc.queueSendSoap(pullMessagesResp(motionMsg("false")), nil)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	s, err := newStream(ctx, fc, Options{
+		PullTimeout:            50 * time.Millisecond,
+		ReconnectAfterFailures: 1,
+		RetryBackoff:           10 * time.Millisecond,
+		InitialTermination:     30 * time.Second,
+	})
+	require.NoError(t, err)
+	defer s.Close()
+
+	ev1 := receive(t, s.Events(), 2*time.Second)
+	assert.True(t, ev1.AfterReconnect, "first event after recreate must carry AfterReconnect=true")
+	assert.Equal(t, StateActive, ev1.State)
+
+	ev2 := receive(t, s.Events(), 2*time.Second)
+	assert.False(t, ev2.AfterReconnect, "subsequent events should not carry AfterReconnect")
+	assert.Equal(t, StateInactive, ev2.State)
+}
+
+// --- Jitter ----------------------------------------------------------
+
+func TestJitter_StaysWithinFraction(t *testing.T) {
+	const base = time.Second
+	low := time.Duration(float64(base) * (1 - jitterFraction))
+	high := time.Duration(float64(base) * (1 + jitterFraction))
+	for i := 0; i < 200; i++ {
+		got := jitter(base)
+		assert.GreaterOrEqual(t, got, low, "iteration %d", i)
+		assert.LessOrEqual(t, got, high, "iteration %d", i)
+	}
+}
+
+func TestJitter_ZeroAndNegativeReturnPositive(t *testing.T) {
+	assert.Greater(t, jitter(0), time.Duration(0))
+	assert.Greater(t, jitter(-time.Second), time.Duration(0))
+}
+
+func TestJitter_VariesAcrossCalls(t *testing.T) {
+	first := jitter(time.Second)
+	allEqual := true
+	for i := 0; i < 10; i++ {
+		if jitter(time.Second) != first {
+			allEqual = false
+			break
+		}
+	}
+	assert.False(t, allEqual, "jitter is producing a constant; rand seed not working")
+}
+
+func TestMaxRecreateBackoff_Is5Minutes(t *testing.T) {
+	assert.Equal(t, 5*time.Minute, maxRecreateBackoff)
 }
